+++ crates/common/types/requests.rs
@@ -50,7 +50,7 @@ impl Serialize for EncodedRequests {
 }
 
 impl RLPEncode for EncodedRequests {
-    fn encode(&self, buf: &mut dyn bytes::BufMut) {
+    fn encode(&self, buf: &mut Vec<u8>) {
         self.0.encode(buf)
     }
 }