+++ crates/networking/p2p/rlpx/eth/status.rs
@@ -6,7 +6,6 @@ use crate::rlpx::{
     error::PeerConnectionError,
     utils::{snappy_compress, snappy_decompress},
 };
-use bytes::BufMut;
 use ethrex_common::types::{BlockHash, ForkId};
 use ethrex_rlp::{
     error::{RLPDecodeError, RLPEncodeError},
@@ -73,7 +72,7 @@ impl StatusDataPost68 {
         })
     }
 
-    pub fn encode(&self, buf: &mut dyn BufMut) -> Result<(), RLPEncodeError> {
+    pub fn encode(&self, buf: &mut Vec<u8>) -> Result<(), RLPEncodeError> {
         let mut encoded_data = vec![];
         Encoder::new(&mut encoded_data)
             .encode_field(&self.eth_version)
@@ -86,7 +85,7 @@ impl StatusDataPost68 {
             .finish();
 
         let msg_data = snappy_compress(encoded_data)?;
-        buf.put_slice(&msg_data);
+        buf.extend_from_slice(&msg_data);
         Ok(())
     }